+++ src/ArduinoJson/Document/DynamicJsonDocument.hpp
@@ -13,16 +13,24 @@ namespace ARDUINOJSON_NAMESPACE {
 class DynamicJsonDocument : public JsonDocument {
  public:
   explicit DynamicJsonDocument(size_t capa)
-      : JsonDocument(allocPool(addPadding(capa))) {}
+      : JsonDocument(allocPool(capa)) {}
 
   DynamicJsonDocument(const DynamicJsonDocument& src)
-      : JsonDocument(allocPool(src.capacity())) {
-    copy(src);
+      : JsonDocument(allocPool(src.memoryUsage())) {
+    set(src);
   }
 
-  DynamicJsonDocument(const JsonDocument& src)
-      : JsonDocument(allocPool(src.capacity())) {
-    copy(src);
+  template <typename T>
+  DynamicJsonDocument(const T& src,
+                      typename enable_if<IsVisitable<T>::value>::type* = 0)
+      : JsonDocument(allocPool(src.memoryUsage())) {
+    set(src);
+  }
+
+  // disambiguate
+  DynamicJsonDocument(VariantRef src)
+      : JsonDocument(allocPool(src.memoryUsage())) {
+    set(src);
   }
 
   ~DynamicJsonDocument() {
@@ -31,19 +39,20 @@ class DynamicJsonDocument : public JsonDocument {
 
   DynamicJsonDocument& operator=(const DynamicJsonDocument& src) {
     reallocPoolIfTooSmall(src.memoryUsage());
-    copy(src);
+    set(src);
     return *this;
   }
 
   template <typename T>
-  DynamicJsonDocument& operator=(const JsonDocument& src) {
+  DynamicJsonDocument& operator=(const T& src) {
     reallocPoolIfTooSmall(src.memoryUsage());
-    copy(src);
+    set(src);
     return *this;
   }
 
  private:
-  MemoryPool allocPool(size_t capa) {
+  MemoryPool allocPool(size_t requiredSize) {
+    size_t capa = addPadding(requiredSize);
     return MemoryPool(reinterpret_cast<char*>(malloc(capa)), capa);
   }